experiment: use request handler for update indicator
